cleaningup api/index.ts endpoints into a logical endpoints dir
