+++ src/api/endpoints/index.ts
@@ -1,0 +1,9 @@
+
+export * from './root.get'
+export * from './upload.post'
+export * from './topic.delete'
+export * from './profile.get'
+export * from './index.get'
+export * from './article.get'
+export * from './search.get'
+export * from './lookup.get'
